+++ pkg/app/validate_app_test.go
@@ -32,6 +32,7 @@ func Test_ValidateApp(t *testing.T) {
 			"giantswarm",
 			"kube-system",
 			"monitoring",
+			"org-giantswarm",
 		},
 		GroupWhitelist: []string{
 			"giantswarm:giantswarm:giantswarm-admins",
@@ -482,6 +483,109 @@ func Test_ValidateApp(t *testing.T) {
 			configMaps: []*corev1.ConfigMap{
 				newTestConfigMap("app-operator-cluster-values", "giantswarm"),
 			},
+		},
+		{
+			// App created by some GS member with access to the clusters, but being in the group
+			// not explicitly listed in the user whitelists. Still, app created in private namespaces
+			// means we pass the validation.
+			name: "referencing protected kubeconfig from a private app",
+			obj: &admissionv1.AdmissionRequest{
+				Operation: "CREATE",
+				Object: runtime.RawExtension{
+					Raw: []byte(`
+						{
+							"apiVersion": "application.giantswarm.io/v1alpha1",
+							"kind": "App",
+							"metadata": {
+    							"name": "fake-controller",
+    							"namespace": "org-giantswarm",
+    							"labels": {
+									"giantswarm.io/cluster": "gsgang",
+									"app-operator.giantswarm.io/version": "0.0.0"
+    							}
+							},
+							"spec": {
+    							"catalog": "control-plane",
+    							"name": "fake-controller",
+    							"namespace": "giantswarm",
+    							"kubeConfig": {
+									"context": {
+										"name": "gsgang-admin@gsgang"
+									},
+									"inCluster": false,
+									"secret": {
+										"name": "gsgang-kubeconfig",
+										"namespace": "org-giantswarm"
+									}
+								},
+								"version": "0.3.0"
+							}
+						}
+					`),
+				},
+				UserInfo: authv1.UserInfo{
+					Username: "user@giantswarm.io",
+					Groups: []string{
+						"customer:giantswarm:Employees",
+					},
+				},
+			},
+			catalogs: []*v1alpha1.Catalog{
+				newTestCatalog("control-plane", "giantswarm"),
+			},
+			secrets: []*corev1.Secret{
+				newTestSecret("gsgang-kubeconfig", "org-giantswarm"),
+			},
+		},
+		{
+			name: "referencing protected kubeconfig as a regular user",
+			obj: &admissionv1.AdmissionRequest{
+				Operation: "CREATE",
+				Object: runtime.RawExtension{
+					Raw: []byte(`
+						{
+							"apiVersion": "application.giantswarm.io/v1alpha1",
+							"kind": "App",
+							"metadata": {
+    							"name": "malicious-app",
+    							"namespace": "default",
+    							"labels": {
+									"app-operator.giantswarm.io/version": "0.0.0"
+    							}
+							},
+							"spec": {
+    							"catalog": "custom",
+    							"name": "malicious-app",
+    							"namespace": "giantswarm",
+    							"kubeConfig": {
+									"context": {
+										"name": "gsgang-admin@gsgang"
+									},
+									"inCluster": false,
+									"secret": {
+										"name": "gsgang-kubeconfig",
+										"namespace": "org-giantswarm"
+									}
+								},
+								"version": "0.3.0"
+							}
+						}
+					`),
+				},
+				UserInfo: authv1.UserInfo{
+					Username: "00001@customer.onmicrosoft.com",
+					Groups: []string{
+						"customer:CUSTOMER-GROUP",
+					},
+				},
+			},
+			catalogs: []*v1alpha1.Catalog{
+				newTestCatalog("custom", "default"),
+			},
+			secrets: []*corev1.Secret{
+				newTestSecret("gsgang-kubeconfig", "org-giantswarm"),
+			},
+			expectedErr: "security violation error: references to `org-giantswarm` namespace not allowed",
 		},
 		{
 			name: "install blacklisted app as regular service account",