+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Extend security validation logic with checking the `kubeconfig` namespace against blacklisted namespaces.
+
 ## [0.17.0] - 2022-05-27
 
 ### Added